Collect channel stats
