+++ ibm_mq/datadog_checks/ibm_mq/collectors/stats_collector.py
@@ -2,18 +2,12 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 from pymqi.CMQC import MQRC_NO_MSG_AVAILABLE
-from pymqi.CMQCFC import (
-    MQCACH_CHANNEL_NAME,
-    MQCMD_STATISTICS_CHANNEL,
-    MQCMD_STATISTICS_MQI,
-    MQCMD_STATISTICS_Q,
-    MQGACF_CHL_STATISTICS_DATA,
-    MQIACH_CHANNEL_TYPE,
-    MQIAMO_MSGS,
-)
+from pymqi.CMQCFC import MQCMD_STATISTICS_CHANNEL, MQCMD_STATISTICS_MQI, MQCMD_STATISTICS_Q
 
-from datadog_checks.base import AgentCheck, to_native_string
+from datadog_checks.base import AgentCheck
 from datadog_checks.ibm_mq.collectors.utils import CustomPCFExecute
+
+from ..stats_wrapper.channel_stats import ChannelStats
 
 try:
     import pymqi
@@ -23,20 +17,6 @@ except ImportError as e:
     pymqi = None
 
 STATISTICS_QUEUE_NAME = 'SYSTEM.ADMIN.STATISTICS.QUEUE'
-
-
-CHANNEL_TYPE_TO_STR = {
-    pymqi.CMQC.MQCHT_SENDER: 'sender',
-    pymqi.CMQC.MQCHT_SERVER: 'server',
-    pymqi.CMQC.MQCHT_RECEIVER: 'receiver',
-    pymqi.CMQC.MQCHT_REQUESTER: 'requester',
-    pymqi.CMQC.MQCHT_CLUSRCVR: 'clusrcvr',
-    pymqi.CMQC.MQCHT_CLUSSDR: 'clussdr',
-}
-
-
-def get_channel_type(raw_type):
-    return CHANNEL_TYPE_TO_STR.get(raw_type, 'unknown')
 
 
 class StatsCollector(object):
@@ -50,26 +30,11 @@ class StatsCollector(object):
 
         try:
             while True:
-                raw_message = queue.get()
-                message, header = CustomPCFExecute.unpack(raw_message)
+                bin_message = queue.get()
+                message, header = CustomPCFExecute.unpack(bin_message)
 
                 if header.Command == MQCMD_STATISTICS_CHANNEL:
-                    channels = message[MQGACF_CHL_STATISTICS_DATA]
-                    self.check.log.info(
-                        {
-                            'type': 'MQCMD_STATISTICS_CHANNEL',
-                            'MQGACF_CHL_STATISTICS_DATA': message[MQGACF_CHL_STATISTICS_DATA],
-                        }
-                    )
-                    for channel_info in channels:
-                        channel_name = to_native_string(channel_info[MQCACH_CHANNEL_NAME]).strip()
-                        channel_type = get_channel_type(channel_info[MQIACH_CHANNEL_TYPE])
-                        tags = [
-                            'channel:{}'.format(channel_name),
-                            'channel_type:{}'.format(channel_type),
-                        ]
-                        self.check.gauge('ibm_mq.stats.channel.msgs', channel_info[MQIAMO_MSGS], tags=tags)
-                        self.check.gauge('ibm_mq.stats.channel.msgs2', channel_info[MQIAMO_MSGS], tags=tags)
+                    self._collect_channel_stats(message)
                 elif header.Command == MQCMD_STATISTICS_MQI:
                     self.check.log.debug('MQCMD_STATISTICS_MQI not implemented yet')
                 elif header.Command == MQCMD_STATISTICS_Q:
@@ -85,3 +50,12 @@ class StatsCollector(object):
                 raise
         finally:
             queue.close()
+
+    def _collect_channel_stats(self, message):
+        channel_stats = ChannelStats(message)
+        for channel_info in channel_stats.channels:
+            tags = [
+                'channel:{}'.format(channel_info.name),
+                'channel_type:{}'.format(channel_info.type),
+            ]
+            self.check.gauge('ibm_mq.stats.channel.msgs', channel_info.msgs, tags=tags)
